split the config
