+++ src/main/java/org/cloudfoundry/promregator/auth/OAuth2XSUAAEnricher.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
-import org.cloudfoundry.promregator.config.OAuth2XSUAAAuthenticationConfiguration;
+import org.cloudfoundry.promregator.config.AbstractOAuth2XSUAAAuthenticationConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -23,7 +23,7 @@ public class OAuth2XSUAAEnricher implements AuthenticationEnricher, Closeable {
 	private final CloseableHttpClient httpClient;
 	private final ClientCredentialsTokenFlow tokenClient;
 
-	OAuth2XSUAAEnricher(OAuth2XSUAAAuthenticationConfiguration config) {
+	OAuth2XSUAAEnricher(AbstractOAuth2XSUAAAuthenticationConfiguration config) {
 		this(config, null);
 	}
 
@@ -31,7 +31,7 @@ public class OAuth2XSUAAEnricher implements AuthenticationEnricher, Closeable {
 	 * @param config The configuration.
 	 * @param tokenClient A token client from outside. This is intended only for unit tests.
 	 */
-	OAuth2XSUAAEnricher(OAuth2XSUAAAuthenticationConfiguration config, ClientCredentialsTokenFlow tokenClient) {
+	OAuth2XSUAAEnricher(AbstractOAuth2XSUAAAuthenticationConfiguration config, ClientCredentialsTokenFlow tokenClient) {
 		super();
 		OAuth2ServiceConfiguration c = new OAuth2ServiceConfig(config);
