+++ src/main/java/org/cloudfoundry/promregator/config/AuthenticatorConfiguration.java
@@ -6,8 +6,9 @@ public class AuthenticatorConfiguration {
 	private BasicAuthenticationConfiguration basic;
 
 	private OAuth2XSUAAAuthenticationConfiguration oauth2xsuaa;
-	
-	
+
+	private OAuth2XSUAACertificateAuthenticationConfiguration oauth2xsuaaCertificate;
+
 	public String getType() {
 		return type;
 	}
@@ -31,5 +32,12 @@ public class AuthenticatorConfiguration {
 	public void setOauth2xsuaa(OAuth2XSUAAAuthenticationConfiguration oauth2xsuaa) {
 		this.oauth2xsuaa = oauth2xsuaa;
 	}
-	
+
+	public OAuth2XSUAACertificateAuthenticationConfiguration getOauth2xsuaaCertificate() {
+		return oauth2xsuaaCertificate;
+	}
+
+	public void setOauth2xsuaaCertificate(OAuth2XSUAACertificateAuthenticationConfiguration oauth2xsuaaCertificate) {
+		this.oauth2xsuaaCertificate = oauth2xsuaaCertificate;
+	}
 }
